+++ internal/handlers/handlers.go
@@ -1,4 +1,4 @@
-package app
+package handlers
 
 import (
 	"encoding/base64"
