+++ internal/router/router.go
@@ -2,12 +2,12 @@ package router
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/khodya/go-urlshortener/internal/app"
+	"github.com/khodya/go-urlshortener/internal/handlers"
 )
 
 func SetupRouter() *gin.Engine {
 	r := gin.Default()
-	r.GET("/:id", app.Unfold)
-	r.POST("/", app.Fold)
+	r.GET("/:id", handlers.Unfold)
+	r.POST("/", handlers.Fold)
 	return r
 }
